requirements.txt
aiohappyeyeballs==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.10.5 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.11" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
django==5.1 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.11" and python_version < "4.0"
flake8==7.1.1 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.8 ; python_version >= "3.11" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.11" and python_version < "4.0"
pycodestyle==2.12.1 ; python_version >= "3.11" and python_version < "4.0"
pyflakes==3.2.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.2.2 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.11" and python_version < "4.0"
jwt==1.3.1
djangorestframework-simplejwt==5.3.1
django-allauth
